+++ crates/proof-of-sql/src/base/arrow/column_arrow_conversions.rs
@@ -7,7 +7,6 @@ use arrow::datatypes::{DataType, Field, TimeUnit as ArrowTimeUnit};
 use proof_of_sql_parser::posql_time::{PoSQLTimeUnit, PoSQLTimeZone};
 
 /// Convert [`ColumnType`] values to some arrow [`DataType`]
-#[cfg(feature = "arrow")]
 impl From<&ColumnType> for DataType {
     fn from(column_type: &ColumnType) -> Self {
         match column_type {
@@ -37,7 +36,6 @@ impl From<&ColumnType> for DataType {
 }
 
 /// Convert arrow [`DataType`] values to some [`ColumnType`]
-#[cfg(feature = "arrow")]
 impl TryFrom<DataType> for ColumnType {
     type Error = String;
 
@@ -70,7 +68,6 @@ impl TryFrom<DataType> for ColumnType {
     }
 }
 /// Convert [`ColumnField`] values to arrow Field
-#[cfg(feature = "arrow")]
 impl From<&ColumnField> for Field {
     fn from(column_field: &ColumnField) -> Self {
         Field::new(
