+++ crates/proof-of-sql/src/base/database/mod.rs
@@ -19,13 +19,6 @@ mod literal_value;
 pub use literal_value::LiteralValue;
 
 mod table_ref;
-#[cfg(feature = "arrow")]
-pub use crate::base::arrow::{
-    arrow_array_to_column_conversion::{ArrayRefExt, ArrowArrayToColumnConversionError},
-    owned_and_arrow_conversions::OwnedArrowConversionError,
-    record_batch_utility::ToArrow,
-    scalar_and_i256_conversions,
-};
 pub use table_ref::TableRef;
 
 mod owned_column;
