+++ crates/proof-of-sql/src/sql/proof/mod.rs
@@ -68,3 +68,6 @@ pub(crate) use result_element_serialization::{
 
 mod first_round_builder;
 pub(crate) use first_round_builder::FirstRoundBuilder;
+
+#[cfg(all(test, feature = "arrow"))]
+mod provable_query_result_test;
