+++ crates/proof-of-sql/src/sql/proof/provable_query_result_test.rs
@@ -1,11 +1,9 @@
-use crate::{
-    base::{
-        database::{Column, ColumnField, ColumnType},
-        math::decimal::Precision,
-        polynomial::compute_evaluation_vector,
-        scalar::{Curve25519Scalar, Scalar},
-    },
-    sql::proof::{ProvableQueryResult, QueryError},
+use super::{ProvableQueryResult, QueryError};
+use crate::base::{
+    database::{Column, ColumnField, ColumnType},
+    math::decimal::Precision,
+    polynomial::compute_evaluation_vector,
+    scalar::{Curve25519Scalar, Scalar},
 };
 use alloc::sync::Arc;
 use arrow::{
